Move EventLevel extension to separate file
Refactored EventLevel typealias and its extension from Event.swift into a new EventLevel.swift file for better code organization and separation of concerns.
