+++ Sources/Scout/UI/Event/Event.swift
@@ -7,7 +7,6 @@
 
 import CloudKit
 import Logging
-import SwiftUI
 
 struct Event: Identifiable {
     let name: String
@@ -46,39 +45,5 @@ extension Event {
         id = record.recordID
         userID = record["user_id"].flatMap(UUID.init)
         sessionID = record["session_id"].flatMap(UUID.init)
-    }
-}
-
-typealias EventLevel = Logger.Level
-
-extension EventLevel {
-    var description: String {
-        switch self {
-        case .notice:
-            "Notice"
-        case .debug:
-            "Debug"
-        case .trace:
-            "Trace"
-        case .info:
-            "Info"
-        case .warning:
-            "Warning"
-        case .error:
-            "Error"
-        case .critical:
-            "Critical"
-        }
-    }
-
-    var color: Color? {
-        switch self {
-        case .notice, .debug, .trace, .info:
-            return nil
-        case .warning, .error:
-            return .yellow
-        case .critical:
-            return .red
-        }
     }
 }
